Redesign onnx pass to enable shape type dependent peephole conversion
