+++ test/onnx/test_pytorch_onnx_onnxruntime.py
@@ -1471,10 +1471,12 @@ class TestONNXRuntime(unittest.TestCase):
         self.run_test(CopyModel(), (x, update))
 
     @skipIfUnsupportedMinOpsetVersion(11)
-    @disableScriptTest()  # Missing input size (with ellipsis indexing)
     def test_copy_ellipsis_tracing(self):
         class CopyModel(torch.nn.Module):
             def forward(self, x, update):
+                # Insert reshape node to ensure no shape/type info for
+                # x in scripting, without onnx shape inference.
+                x = x.reshape(4, 3, 5, 6)
                 x[2, ..., 1:3] = update
                 return x
 
@@ -1745,7 +1747,6 @@ class TestONNXRuntime(unittest.TestCase):
         x = torch.randn(4, 6, 180, 180)
         self.run_test(model, x)
 
-    @disableScriptTest()
     def test_groupnorm_noaffine(self):
         model = torch.nn.GroupNorm(4, 8, 0.002, affine=False)
         x = torch.randn(3, 8, 224, 224)
